+++ packages/core/env/authconfig/run.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+lxc exec authconfig bash <<EOF
+cd /nikita/packages/core
+npm test
+EOF